Divide CFG class (though no luck actually splitting it)
