+++ src/StudioCore/CFG.cs
@@ -17,7 +17,7 @@ internal partial class CfgSerializerContext : JsonSerializerContext
 {
 }
 
-public class CFG
+public partial class CFG
 {
     public const string FolderName = "DSMapStudio";
     public const string Config_FileName = "DSMapStudio_Config.json";
@@ -27,9 +27,6 @@ public class CFG
     public static bool IsEnabled = true;
 
     private static readonly object _lock_SaveLoadCFG = new();
-
-    //private string _Param_Export_Array_Delimiter = "|";
-    private string _Param_Export_Delimiter = ",";
 
     // JsonExtensionData stores info in config file not present in class in order to retain settings between versions.
 #pragma warning disable IDE0051
@@ -65,139 +62,7 @@ public class CFG
     public bool MapAliases_ShowTagsInBrowser = true;
     public bool MapAliases_ShowAliasAddition = false;
 
-    // Settings: Map Editor
-    public bool Viewport_Enable_Selection_Outline = false;
-
-    public bool MapEditor_MapObjectList_ShowMapNames = true;
-    public bool MapEditor_MapObjectList_ShowCharacterNames = true;
-    public bool MapEditor_MapObjectList_ShowAssetNames = true;
-    public bool MapEditor_MapObjectList_ShowMapPieceNames = true;
-    public bool MapEditor_MapObjectList_ShowPlayerCharacterNames = true;
-    public bool MapEditor_MapObjectList_ShowSystemCharacterNames = true;
-    public bool MapEditor_MapObjectList_ShowTreasureNames = true;
-
-    public bool EnableFrustrumCulling = false;
-    public bool Map_AlwaysListLoadedMaps = true;
-    public bool EnableEldenRingAutoMapOffset = true;
-
-    public bool Map_EnableViewportGrid = false;
-    public int Map_ViewportGridType = 0;
-    public Vector3 GFX_Viewport_Grid_Color = Utils.GetDecimalColor(Color.Red);
-    public int Map_ViewportGrid_TotalSize = 1000;
-    public int Map_ViewportGrid_IncrementSize = 10;
-
-    public float Map_ViewportGrid_Offset = 0;
-
-    public float Map_ViewportGrid_ShortcutIncrement = 1;
-
-    public float Map_MoveSelectionToCamera_Radius = 3.0f;
-    public float GFX_Camera_FOV { get; set; } = 60.0f;
-    public float GFX_Camera_MoveSpeed_Slow { get; set; } = 1.0f;
-    public float GFX_Camera_MoveSpeed_Normal { get; set; } = 20.0f;
-    public float GFX_Camera_MoveSpeed_Fast { get; set; } = 200.0f;
-    public float GFX_Camera_Sensitivity { get; set; } = 0.0160f;
-    public float GFX_RenderDistance_Max { get; set; } = 50000.0f;
-    public float Map_ArbitraryRotation_X_Shift { get; set; } = 90.0f;
-    public float Map_ArbitraryRotation_Y_Shift { get; set; } = 90.0f;
-
-    public float GFX_Framerate_Limit_Unfocused = 20.0f;
-    public float GFX_Framerate_Limit = 60.0f;
-    public uint GFX_Limit_Buffer_Flver_Bone = 65536;
-    public uint GFX_Limit_Buffer_Indirect_Draw = 50000;
-    public int GFX_Limit_Renderables = 50000;
-
-    public float GFX_Wireframe_Color_Variance = 0.11f;
-
-    public Vector3 GFX_Renderable_Box_BaseColor = Utils.GetDecimalColor(Color.Blue);
-    public Vector3 GFX_Renderable_Box_HighlightColor = Utils.GetDecimalColor(Color.DarkViolet);
-
-    public Vector3 GFX_Renderable_Cylinder_BaseColor = Utils.GetDecimalColor(Color.Blue);
-    public Vector3 GFX_Renderable_Cylinder_HighlightColor = Utils.GetDecimalColor(Color.DarkViolet);
-
-    public Vector3 GFX_Renderable_Sphere_BaseColor = Utils.GetDecimalColor(Color.Blue);
-    public Vector3 GFX_Renderable_Sphere_HighlightColor = Utils.GetDecimalColor(Color.DarkViolet);
-
-    public Vector3 GFX_Renderable_Point_BaseColor = Utils.GetDecimalColor(Color.Yellow);
-    public Vector3 GFX_Renderable_Point_HighlightColor = Utils.GetDecimalColor(Color.DarkViolet);
-
-    public Vector3 GFX_Renderable_DummyPoly_BaseColor = Utils.GetDecimalColor(Color.Yellow);
-    public Vector3 GFX_Renderable_DummyPoly_HighlightColor = Utils.GetDecimalColor(Color.DarkViolet);
-
-    public Vector3 GFX_Renderable_BonePoint_BaseColor = Utils.GetDecimalColor(Color.Blue);
-    public Vector3 GFX_Renderable_BonePoint_HighlightColor = Utils.GetDecimalColor(Color.DarkViolet);
-
-    public Vector3 GFX_Renderable_ModelMarker_Chr_BaseColor = Utils.GetDecimalColor(Color.Firebrick);
-    public Vector3 GFX_Renderable_ModelMarker_Chr_HighlightColor = Utils.GetDecimalColor(Color.Tomato);
-
-    public Vector3 GFX_Renderable_ModelMarker_Object_BaseColor = Utils.GetDecimalColor(Color.MediumVioletRed);
-    public Vector3 GFX_Renderable_ModelMarker_Object_HighlightColor = Utils.GetDecimalColor(Color.DeepPink);
-
-    public Vector3 GFX_Renderable_ModelMarker_Player_BaseColor = Utils.GetDecimalColor(Color.DarkOliveGreen);
-    public Vector3 GFX_Renderable_ModelMarker_Player_HighlightColor = Utils.GetDecimalColor(Color.OliveDrab);
-
-    public Vector3 GFX_Renderable_ModelMarker_Other_BaseColor = Utils.GetDecimalColor(Color.Wheat);
-    public Vector3 GFX_Renderable_ModelMarker_Other_HighlightColor = Utils.GetDecimalColor(Color.AntiqueWhite);
-
-    public Vector3 GFX_Renderable_PointLight_BaseColor = Utils.GetDecimalColor(Color.YellowGreen);
-    public Vector3 GFX_Renderable_PointLight_HighlightColor = Utils.GetDecimalColor(Color.Yellow);
-
-    public Vector3 GFX_Renderable_SpotLight_BaseColor = Utils.GetDecimalColor(Color.Goldenrod);
-    public Vector3 GFX_Renderable_SpotLight_HighlightColor = Utils.GetDecimalColor(Color.Violet);
-
-    public Vector3 GFX_Renderable_DirectionalLight_BaseColor = Utils.GetDecimalColor(Color.Cyan);
-    public Vector3 GFX_Renderable_DirectionalLight_HighlightColor = Utils.GetDecimalColor(Color.AliceBlue);
-
-    public Vector3 GFX_Gizmo_X_BaseColor = new(0.952f, 0.211f, 0.325f);
-    public Vector3 GFX_Gizmo_X_HighlightColor = new(1.0f, 0.4f, 0.513f);
-
-    public Vector3 GFX_Gizmo_Y_BaseColor = new(0.525f, 0.784f, 0.082f);
-    public Vector3 GFX_Gizmo_Y_HighlightColor = new(0.713f, 0.972f, 0.270f);
-
-    public Vector3 GFX_Gizmo_Z_BaseColor = new(0.219f, 0.564f, 0.929f);
-    public Vector3 GFX_Gizmo_Z_HighlightColor = new(0.407f, 0.690f, 1.0f);
-    public RenderFilter LastSceneFilter { get; set; } = RenderFilter.All ^ RenderFilter.Light;
-
-    public RenderFilterPreset SceneFilter_Preset_01 { get; set; } = new("Map",
-        RenderFilter.MapPiece | RenderFilter.Object | RenderFilter.Character | RenderFilter.Region);
-
-    public RenderFilterPreset SceneFilter_Preset_02 { get; set; } = new("Collision",
-        RenderFilter.Collision | RenderFilter.Object | RenderFilter.Character | RenderFilter.Region);
-
-    public RenderFilterPreset SceneFilter_Preset_03 { get; set; } = new("Collision & Navmesh",
-        RenderFilter.Collision | RenderFilter.Navmesh | RenderFilter.Object | RenderFilter.Character |
-        RenderFilter.Region);
-
-    public RenderFilterPreset SceneFilter_Preset_04 { get; set; } = new("Lighting (Map)",
-        RenderFilter.MapPiece | RenderFilter.Object | RenderFilter.Character | RenderFilter.Light);
-
-    public RenderFilterPreset SceneFilter_Preset_05 { get; set; } = new("Lighting (Collision)",
-        RenderFilter.Collision | RenderFilter.Object | RenderFilter.Character | RenderFilter.Light);
-
-    public RenderFilterPreset SceneFilter_Preset_06 { get; set; } = new("All", RenderFilter.All);
-
     // Settings: Model Editor
-
-    // Settings: Param Editor
-    public bool Param_AdvancedMassedit = false;
-    public bool Param_AllowFieldReorder = true;
-    public bool Param_AlphabeticalParams = true;
-    public bool Param_DisableLineWrapping = false;
-    public bool Param_DisableRowGrouping = false;
-    public bool Param_HideEnums = false;
-    public bool Param_HideReferenceRows = false;
-    public bool Param_MakeMetaNamesPrimary = true;
-    public bool Param_PasteAfterSelection = false;
-    public bool Param_PasteThenSelect = true;
-    public bool Param_ShowFieldOffsets = false;
-    public bool Param_ShowHotkeysInContextMenu = true;
-    public bool Param_ShowSecondaryNames = true;
-    public bool Param_ShowVanillaParams = true;
-    public bool UI_CompactParams = false;
-
-    // Settings: Text Editor
-    public bool FMG_NoFmgPatching = false;
-    public bool FMG_NoGroupedFmgEntries = false;
-    public bool FMG_ShowOriginalNames = false;
 
     // CFG
     public static CFG Current { get; private set; }
@@ -213,26 +78,6 @@ public class CFG
 
     public int GFX_Display_X { get; set; } = 0;
     public int GFX_Display_Y { get; set; } = 23;
-
-    public string Param_Export_Delimiter
-    {
-        get
-        {
-            if (_Param_Export_Delimiter.Length == 0)
-            {
-                _Param_Export_Delimiter = Default.Param_Export_Delimiter;
-            }
-            else if (_Param_Export_Delimiter == "|")
-            {
-                _Param_Export_Delimiter =
-                    Default
-                        .Param_Export_Delimiter; // Temporary measure to prevent conflicts with byte array delimiters. Will be removed later.
-            }
-
-            return _Param_Export_Delimiter;
-        }
-        set => _Param_Export_Delimiter = value;
-    }
 
     public static string GetConfigFilePath()
     {
@@ -427,22 +272,5 @@ public class CFG
             }
             return false;
         }
-    }
-
-    public class RenderFilterPreset
-    {
-        [JsonConstructor]
-        public RenderFilterPreset()
-        {
-        }
-
-        public RenderFilterPreset(string name, RenderFilter filters)
-        {
-            Name = name;
-            Filters = filters;
-        }
-
-        public string Name { get; set; }
-        public RenderFilter Filters { get; set; }
     }
 }
