+++ src/StudioCore/ParamEditor/ParamCFG.cs
@@ -1,0 +1,53 @@
+﻿using StudioCore.Platform;
+using StudioCore.Scene;
+using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.IO;
+using System.Numerics;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+
+namespace StudioCore;
+
+public partial class CFG
+{
+
+    //private string _Param_Export_Array_Delimiter = "|";
+    private string _Param_Export_Delimiter = ",";
+
+    public bool Param_AdvancedMassedit = false;
+    public bool Param_AllowFieldReorder = true;
+    public bool Param_AlphabeticalParams = true;
+    public bool Param_DisableLineWrapping = false;
+    public bool Param_DisableRowGrouping = false;
+    public bool Param_HideEnums = false;
+    public bool Param_HideReferenceRows = false;
+    public bool Param_MakeMetaNamesPrimary = true;
+    public bool Param_PasteAfterSelection = false;
+    public bool Param_PasteThenSelect = true;
+    public bool Param_ShowFieldOffsets = false;
+    public bool Param_ShowHotkeysInContextMenu = true;
+    public bool Param_ShowSecondaryNames = true;
+    public bool Param_ShowVanillaParams = true;
+    public bool UI_CompactParams = false;
+    
+    public string Param_Export_Delimiter
+    {
+        get
+        {
+            if (_Param_Export_Delimiter.Length == 0)
+            {
+                _Param_Export_Delimiter = Default.Param_Export_Delimiter;
+            }
+            else if (_Param_Export_Delimiter == "|")
+            {
+                _Param_Export_Delimiter =
+                    Default.Param_Export_Delimiter; // Temporary measure to prevent conflicts with byte array delimiters. Will be removed later.
+            }
+
+            return _Param_Export_Delimiter;
+        }
+        set => _Param_Export_Delimiter = value;
+    }
+}
