+++ src/StudioCore/TextEditor/TextCFG.cs
@@ -1,0 +1,17 @@
+﻿using StudioCore.Platform;
+using StudioCore.Scene;
+using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.IO;
+using System.Numerics;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+
+namespace StudioCore;
+public partial class CFG
+{
+    public bool FMG_NoFmgPatching = false;
+    public bool FMG_NoGroupedFmgEntries = false;
+    public bool FMG_ShowOriginalNames = false;
+}
